requirements.txt
anyio==3.7.1
azure-common==1.1.28
azure-core==1.28.0
azure-identity==1.13.0
azure-keyvault==4.2.0
azure-keyvault-certificates==4.7.0
azure-keyvault-keys==4.8.0
azure-keyvault-secrets==4.7.0
black==23.7.0
Brotli==1.0.9
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.5
colorama==0.4.6
cryptography==41.0.2
dynaconf==3.2.0
FlightRadarAPI==1.3.10
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
isodate==0.6.1
msal==1.22.0
msal-extensions==1.0.0
mypy-extensions==1.0.0
numpy==1.25.1
packaging==23.1
pandas==2.0.3
pathspec==0.11.1
platformdirs==3.9.1
plotly==5.15.0
portalocker==2.7.0
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-telegram-bot==20.4
pytz==2023.3
requests==2.31.0
six==1.16.0
sniffio==1.3.0
tenacity==8.2.2
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
